# pkg/docs/swagger.yaml
info:
  contact: {}
  title: Knowledge API
paths:
  /datasets:
    get:
      description: List all datasets
      produces:
      - application/json
      responses: {}
      summary: List all datasets
      tags:
      - datasets
  /datasets/{id}:
    delete:
      description: Delete a dataset by ID
      parameters:
      - description: Dataset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete a dataset
      tags:
      - datasets
    get:
      description: Get a dataset by ID
      parameters:
      - description: Dataset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get a dataset
      tags:
      - datasets
  /datasets/{id}/documents/{doc_id}:
    delete:
      consumes:
      - application/json
      description: Remove a document from a dataset by ID
      parameters:
      - description: Dataset ID
        in: path
        name: id
        required: true
        type: string
      - description: Document ID
        in: path
        name: doc_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Remove a document from a dataset
      tags:
      - datasets
  /datasets/{id}/files/{file_id}:
    delete:
      consumes:
      - application/json
      description: Remove a file from a dataset by ID
      parameters:
      - description: Dataset ID
        in: path
        name: id
        required: true
        type: string
      - description: File ID
        in: path
        name: file_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Remove a file from a dataset
      tags:
      - datasets
  /datasets/{id}/ingest:
    post:
      consumes:
      - application/json
      description: Ingest content into a dataset by ID
      parameters:
      - description: Dataset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Ingest content into a dataset
      tags:
      - datasets
  /datasets/{id}/query:
    post:
      consumes:
      - application/json
      description: Query a dataset by ID
      parameters:
      - description: Dataset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Query a dataset
      tags:
      - datasets
  /datasets/{id}/retrieve:
    post:
      consumes:
      - application/json
      description: Retrieve content from a dataset by ID
      parameters:
      - description: Dataset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Retrieve content from a dataset
      tags:
      - datasets
  /datasets/create:
    post:
      consumes:
      - application/json
      description: Create a new dataset
      produces:
      - application/json
      responses: {}
      summary: Create a new dataset
      tags:
      - datasets
swagger: "2.0"